Add multinode support
Support for a mutlinode riak_core release.  It borrows ideas and code
from Rebar's simplenode template and Riak.  It assumes that the user
wants to create both an application and a release based on that
application.  May want to split those two responsibilities apart in
the future but for now I just want to get something working.